% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{add_bayes}
\alias{add_bayes}
\title{Addition in Bayes space}
\usage{
add_bayes(tensor1, tensor2, tensor_ref, do_projection = FALSE)
}
\arguments{
\item{tensor1}{A \code{3x3} symmetric matrix.}

\item{tensor2}{A \code{3x3} symmetric matrix.}

\item{tensor_ref}{The SDP tensor associated to the reference measure of the
Bayes space.}

\item{do_projection}{A boolean specifying whether the result of the operation
should be projected onto the space of diffusion signals (default not to).}
}
\value{
A \code{3x3} symmetric matrix.
}
\description{
Addition in Bayes space
}
\examples{
D1 <- diag(c(1.7, 0.3, 0.1)) * 1e-3
D2 <- diag(c(0.3, 1.7, 0.1)) * 1e-3
Dr <- diag(c(1.7, 0.3, 0.1)) * 1e-3
add_bayes(D1, D2, Dr)
}
