% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_tensor}
\alias{as_tensor}
\title{Vector-To-Tensor Representation}
\usage{
as_tensor(vector)
}
\arguments{
\item{vector}{A vector of size 6 storing the 6 unique components of a tensor.}
}
\value{
A 3x3 symmetric postive definite matrix as output tensor.
}
\description{
Vector-To-Tensor Representation
}
\examples{
v <- seq_len(6L)
as_tensor(v)
}
