% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_vector}
\alias{as_vector}
\title{Tensor-To-Vector Representation}
\usage{
as_vector(tensor)
}
\arguments{
\item{tensor}{A 3x3 symmetric postive definite matrix.}
}
\value{
A vector of size 6 storing the 6 unique components of the input tensor.
}
\description{
Tensor-To-Vector Representation
}
\examples{
D <- cbind(c(1, 2, 4), c(2, 3, 5), c(4, 5, 6))
as_vector(D)
}
