% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_add}
\alias{bayes_add}
\title{Addition in Bayes space}
\usage{
bayes_add(isigma1, isigma2, neutral_variance = neutral_element)
}
\arguments{
\item{isigma1}{A \code{3x3} symmetric positive definite matrix.}

\item{isigma2}{A \code{3x3} symmetric positive definite matrix.}

\item{neutral_variance}{The neutral element of the Bayes space (default: 1e5).}
}
\value{
A \code{3x3} symmetric positive definite matrix.
}
\description{
Addition in Bayes space
}
\examples{
D1 <- diag(c(1.7, 0.3, 0.1)) * 1e-3
D2 <- diag(c(0.3, 1.7, 0.1)) * 1e-3
bayes_add(D1, D2)
}
