% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{bayes_distance}
\alias{bayes_distance}
\title{Distance in Bayes space}
\usage{
bayes_distance(isigma1, isigma2, neutral_variance = neutral_element)
}
\arguments{
\item{isigma1}{A \code{3x3} symmetric positive definite matrix.}

\item{isigma2}{A \code{3x3} symmetric positive definite matrix.}

\item{neutral_variance}{The neutral element of the Bayes space (default:
1e5).}
}
\value{
A positive scalar measuring the distance between the two input
  Gaussians in Bayes geometry.
}
\description{
Distance in Bayes space
}
\examples{
D1 <- diag(c(1.7, 0.3, 0.1)) * 1e-3
D2 <- diag(c(0.3, 1.7, 0.1)) * 1e-3
bayes_distance(D1, D2)
}
