% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_mean}
\alias{bayes_mean}
\title{Mean element in Bayes space}
\usage{
bayes_mean(isigmaList, neutral_variance = neutral_element)
}
\arguments{
\item{isigmaList}{A list of \code{3x3} symmetric positive definite matrices.}

\item{neutral_variance}{The neutral element of the Bayes space (default: 1e5).}
}
\value{
A \code{3x3} symmetric positive definite matrix.
}
\description{
Mean element in Bayes space
}
\examples{
D <- list(D1 = diag(c(1.7, 0.3, 0.1)) * 1e-3,
          D2 = diag(c(0.3, 1.7, 0.1)) * 1e-3)
bayes_mean(D)
}
