% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes.R
\name{bayes_multiply}
\alias{bayes_multiply}
\title{Scalar multiplication in Bayes space}
\usage{
bayes_multiply(isigma, alpha, neutral_variance = neutral_element)
}
\arguments{
\item{isigma}{A \code{3x3} symmetric positive definite matrix.}

\item{alpha}{A scalar.}

\item{neutral_variance}{The neutral element of the Bayes space (default: 1e5).}
}
\value{
A \code{3x3} symmetric positive definite matrix.
}
\description{
Scalar multiplication in Bayes space
}
\examples{
D <- diag(c(1.7, 0.3, 0.1)) * 1e-3
bayes_multiply(D, 2)
bayes_add(D, D)
}
