\name{BayesR-package}
\alias{BayesR-package}
\alias{BayesR}
\docType{package}

\title{
  Bayesian Regression in R
}

\description{
  R infrastructures for Bayesian regression models.
}

\details{
  \tabular{ll}{
    Package: \tab BayesR\cr
    Type: \tab Package\cr
    Version: \tab 0.1-1\cr
    Date: \tab 2013-04-08\cr
    License: \tab GPL-2 | GPL-3\cr
  }
}

\author{
  Nikolaus Umlauf, Achim Zeileis.
}

\references{
  ...
}


\keyword{package}

\seealso{
  \code{\link{bayesr}}
}

\examples{
## See the documentation of ?bayesr
}

