% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{det_matrix}
\alias{det_matrix}
\title{Determinant of a symmetric matrix}
\usage{
det_matrix(x)
}
\arguments{
\item{x}{A symmetric matrix.}
}
\value{
The determinant of the input symmetric matrix.
}
\description{
Determinant of a symmetric matrix
}
\examples{
M <- matrix(rnorm(9), 3, 3)
det_matrix(M)
}
