% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{dist_bayes}
\alias{dist_bayes}
\title{Distance in Bayes space}
\usage{
dist_bayes(tensor1, tensor2, iR)
}
\arguments{
\item{tensor1}{A \code{3x3} symmetric matrix.}

\item{tensor2}{A \code{3x3} symmetric matrix.}

\item{iR}{The inverse of the SDP tensor associated to the reference measure
of the Bayes space.}
}
\value{
A positive scalar.
}
\description{
Distance in Bayes space
}
\examples{
D1 <- diag(c(1.7, 0.3, 0.1)) * 1e-3
D2 <- diag(c(0.3, 1.7, 0.1)) * 1e-3
dist_bayes(D1, D2)
}
