% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2vec}
\alias{mat2vec}
\title{Tensor-To-Vector Representation}
\usage{
mat2vec(tensor, twice = FALSE)
}
\arguments{
\item{tensor}{A 3x3 symmetric postive definite matrix.}

\item{twice}{A boolean that says whether off-diagonal tensor elements were
doubled in vector representation (default: \code{FALSE}).}
}
\value{
A vector of size 6 storing the 6 unique components of the input tensor.
}
\description{
Tensor-To-Vector Representation
}
\examples{
D <- cbind(c(1, 2, 4), c(2, 3, 5), c(4, 5, 6))
mat2vec(D)
}
