% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{mean_bayes}
\alias{mean_bayes}
\title{Mean element in Bayes space}
\usage{
mean_bayes(tensor_list, weights = uniform_weights(tensor_list),
  shrinkage = 0, tensor_ref, do_projection = FALSE)
}
\arguments{
\item{tensor_list}{A list of \code{3x3} symmetric matrices.}

\item{weights}{A numeric vector specifying the weight of each observation
(default: equiprobability).}

\item{shrinkage}{A scalar between 0 and 1 specifying how much do we want to
shrink the arithmetic mean towards the reference tensor (default: 0).}

\item{tensor_ref}{The SDP tensor associated to the reference measure of the
Bayes space.}

\item{do_projection}{A boolean specifying whether the result of the operation
should be projected onto the space of diffusion signals (default not to).}
}
\value{
A \code{3x3} symmetric matrix.
}
\description{
Mean element in Bayes space
}
\examples{
D <- list(D1 = diag(c(1.7, 0.3, 0.1)) * 1e-3,
          D2 = diag(c(0.3, 1.7, 0.1)) * 1e-3)
mean_bayes(D)
}
