% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{microstructure_distance}
\alias{microstructure_distance}
\title{Distances on brain microstructure}
\usage{
microstructure_distance(tensor1, tensor2)
}
\arguments{
\item{tensor1}{A symmetric definite positive 3x3 matrix.}

\item{tensor2}{A symmetric definite positive 3x3 matrix.}
}
\value{
A list of distances for microstructure parameters.
}
\description{
Distances on brain microstructure
}
\examples{
D1 <- diag(c(1.7, 0.3, 0.1)) * 1e-3
D2 <- diag(c(0.3, 1.7, 0.1)) * 1e-3
microstructure_distance(D1, D2)
}
