% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{multiply_bayes}
\alias{multiply_bayes}
\title{Scalar multiplication in Bayes space}
\usage{
multiply_bayes(tensor, alpha, tensor_ref, do_projection = FALSE)
}
\arguments{
\item{tensor}{A \code{3x3} symmetric matrix.}

\item{alpha}{A scalar.}

\item{tensor_ref}{The SDP tensor associated to the reference measure of the
Bayes space.}

\item{do_projection}{A boolean specifying whether the result of the operation
should be projected onto the space of diffusion signals (default not to).}
}
\value{
A \code{3x3} symmetric matrix.
}
\description{
Scalar multiplication in Bayes space
}
\examples{
D <- diag(c(1.7, 0.3, 0.1)) * 1e-3
multiply_bayes(D, 2)
}
