% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{project}
\alias{project}
\title{Orthogonal projection onto diffusion signal space}
\usage{
project(x, reference_tensor, distance = "bayes")
}
\arguments{
\item{x}{A symmetric 3x3 matrix resulting from a series of operations in
Bayes space.}

\item{reference_tensor}{The SDP tensor defining the reference measure of the
Bayes space.}

\item{distance}{A string indicating whether the projection should be done in
Bayes geometry (default) or in Euclidean geometry.}
}
\value{
An SDP tensor.
}
\description{
Orthogonal projection onto diffusion signal space
}
\examples{
R <- diag(c(1.7, 0.3, 0.1), 3)
Rn <- matrix(mvtnorm::rmvnorm(1, c(R)), 3, 3)
det_matrix(Rn)
Rfro <- project(Rn, R, distance = "frobenius")
Rbay <- project(Rn, R, distance = "bayes")
eigen(Rn, TRUE, TRUE)$values
eigen(Rfro, TRUE, TRUE)$values
eigen(Rbay, TRUE, TRUE)$values
}
