% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmrice}
\alias{rmrice}
\title{MRI Noise Generator}
\usage{
rmrice(tensor, n = 1L, rho = 0.1)
}
\arguments{
\item{tensor}{A symmetric definite positive 3x3 matrix.}

\item{n}{Number of noisy independent versions of the input tensor to be
generated (default: 1).}

\item{rho}{Relative standard deviation of the MRI noise (default: 0.1).}
}
\value{
A list of \code{n} noisy independent
  versions of the input tensor.
}
\description{
MRI Noise Generator
}
\examples{
D <- diag(c(1.7, 0.3, 0.1)) * 1e-3
rmrice(D, 5L, 0.05)
}
