% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-geometry.R
\name{bayes_mean}
\alias{bayes_mean}
\title{Mean element in Bayes space}
\usage{
bayes_mean(isigmaList, weights = rep(1/length(isigmaList),
  length(isigmaList)), lambda = 1, neutral = neutral_element)
}
\arguments{
\item{isigmaList}{A list of \code{3x3} symmetric positive definite matrices.}

\item{weights}{A numeric vector specifying the weight of each observation
(default: equiprobability).}

\item{lambda}{A scalar between 0 and 1 specifying how much trust do we have
in the reference measure, which is assigned a weight of \code{1 - lambda}
(default: 1).}

\item{neutral}{The neutral element of the Bayes space (default:
\code{diag(3e-3, 3L)}.}
}
\value{
A \code{3x3} symmetric positive definite matrix.
}
\description{
Mean element in Bayes space
}
\examples{
D <- list(D1 = diag(c(1.7, 0.3, 0.1)) * 1e-3,
          D2 = diag(c(0.3, 1.7, 0.1)) * 1e-3)
bayes_mean(D)
}
